+++ simple_history/manager.py
@@ -1,6 +1,7 @@
 from __future__ import unicode_literals
 
 from django.db import models
+from django.utils.timezone import now
 
 
 class HistoryDescriptor(object):
@@ -88,3 +89,19 @@ class HistoryManager(models.Manager):
                               history_type='-').exists():
                 continue
             yield last_change.instance
+
+    def bulk_history_create(self, objs, batch_size=None):
+        """Bulk create the history for the objects specified by objs"""
+
+        historical_instances = [
+            self.model(
+                history_date=getattr(instance, '_history_date', now()),
+                history_user=getattr(instance, '_history_user', None),
+                **{
+                    field.attname: getattr(instance, field.attname)
+                    for field in instance._meta.fields
+                }
+            ) for instance in objs]
+
+        return self.model.objects.bulk_create(historical_instances,
+                                              batch_size=batch_size)