+++ CHANGES.rst
@@ -1,6 +1,10 @@
 Changes
 =======
 
+Unreleased
+----------
+- Fixed out-of-memory exception when running populate_history management command (gh-408)
+
 2.1.0 (2018-06-04)
 ------------------
 - Add ability to specify custom history_reason field (gh-379)